move render jinja view to core
