+++ nautobot/extras/api/serializers.py
@@ -1163,17 +1163,3 @@ class WebhookSerializer(ValidatedModelSerializer, NotesSerializerMixin):
             raise serializers.ValidationError(conflicts)
 
         return validated_data
-
-
-#
-# Tools
-#
-
-
-class RenderJinjaSerializer(serializers.Serializer):  # pylint: disable=abstract-method
-    """Serializer for RenderJinjaView."""
-
-    template_code = serializers.CharField(required=True)
-    context = serializers.DictField(default=dict)
-    rendered_template = serializers.CharField(read_only=True)
-    rendered_template_lines = serializers.ListField(read_only=True, child=serializers.CharField())
