+++ nautobot/extras/api/urls.py
@@ -1,5 +1,3 @@
-from django.urls import path
-
 from nautobot.core.api.routers import OrderedDefaultRouter
 
 from . import views
@@ -118,10 +116,4 @@ router.register("webhooks", views.WebhooksViewSet)
 
 
 app_name = "extras-api"
-
-urlpatterns = [
-    # Tools
-    path("render-jinja-template", views.RenderJinjaView.as_view(), name="render_jinja_template"),
-]
-
-urlpatterns += router.urls
+urlpatterns = router.urls
